Use the Slack Web API and re-organise Slack commands
Introduced new models for speaking to the Slack Web API. The RTM api client used a local store, which meant that channel membership was not always up to date when running a new pool.
By calling the Web API directly when required, we can ensure that we always have the current channel membership.