Improve Snooze and Resume commands
Makes the behaviour of Snooze/Resume commands more consistent.
Now, using `snooze all` or `resume all` in a DM with Pearbot will
pause/reactivate pairings in all Pearbot channel a Participant belongs
to. Commands `snooze me` and `resume me` won't work in a direct
conversation with Pearbot but can still be used in a specific channel.